fixing some compilation errors,changing defs.h

+++ include/common/defs.h
@@ -147,9 +147,9 @@ namespace gridtools{  namespace enumtype{
 
 #ifdef CXX11_ENABLED
     using int_t          = int;
-    using short_t        = int;
-    using uint_t         = long unsigned int;
-    using ushort_t       = unsigned int;
+    using short_t        = short;
+    using uint_t         = unsigned int;
+    using ushort_t       = unsigned short;
     template<int_t N>
     using  static_int=boost::mpl::integral_c<int_t,N>;
     template<uint_t N>
@@ -159,10 +159,10 @@ namespace gridtools{  namespace enumtype{
     template<ushort_t N>
     using  static_ushort=boost::mpl::integral_c<ushort_t,N>;
 #else
-    typedef long int                     int_t;
-    typedef int                     short_t;
+    typedef int                     int_t;
+    typedef short                     short_t;
     typedef unsigned int                     uint_t;
-    typedef unsigned int                     ushort_t;
+    typedef unsigned short                     ushort_t;
     template<int_t N>
     struct static_int : boost::mpl::integral_c<int_t,N>{
         typedef boost::mpl::integral_c<int_t,N> type;
